refactor(frontend): hoist FlatSection to shared components (ATT-359 P4)
FlatSection is now used by Account, Resource detail, Resource usage, and
Groups - it outgrew apps/frontend/src/app/resources/details. Move it to
apps/frontend/src/components alongside other cross-page primitives and
refresh the file header.
